requirements.txt
streamlit==1.28.1
tweepy==4.14.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
sqlite3
datetime